fix
Add:
class RealMoney - realize whole and fractional part of money
class InvalidPercentagesBanksException

Changed:
names of some variables
working with money is now in the account
another representation of accounts in class Client

+++ Banks/Services/AccountTransaction.cs
@@ -2,7 +2,7 @@
 {
     public class AccountTransaction
     {
-        public AccountTransaction(ITransaction transaction, IAccount outAccount, IAccount toAccount, long money)
+        public AccountTransaction(ITransaction transaction, IAccount outAccount, IAccount toAccount, float money)
         {
             Transaction = transaction;
 
@@ -17,6 +17,6 @@
         public IAccount OutAccount { get; }
         public IAccount ToAccount { get; }
 
-        public long Money { get; }
+        public float Money { get; }
     }
 }
